chatting 기본 채팅

See merge request swm-13-main/13_swm74/modak-flutter-app!12

+++ lib/constant/font.dart
@@ -10,13 +10,13 @@ class Font {
   static const FontWeight weight_regular = FontWeight.w400;
 
   /// font sizes
-  static const int size_h1 = 26;
-  static const int size_h2 = 24;
-  static const int size_h3 = 22;
-  static const int size_h4 = 20;
-  static const int size_subTitle = 18;
-  static const int size_largeText = 16;
-  static const int size_mediumText = 14;
-  static const int size_smallText = 12;
-  static const int size_caption = 10;
+  static const double size_h1 = 26;
+  static const double size_h2 = 24;
+  static const double size_h3 = 22;
+  static const double size_h4 = 20;
+  static const double size_subTitle = 18;
+  static const double size_largeText = 16;
+  static const double size_mediumText = 14;
+  static const double size_smallText = 12;
+  static const double size_caption = 10;
 }
